requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
openai==1.3.0
Pillow==10.1.0
numpy==1.26.2
python-dotenv==1.0.0
